+++ test/support/person.rb
@@ -1,4 +1,6 @@
 class Person < ActiveRecord::Base
   has_many :books, -> { order(:title) }, as: :author
   has_many :characters, -> { order(:last_name) }, through: :books
+
+  has_many :books_that_start_with_the, -> { order(:title) }, as: :author, class_name: "BookWithDefaultScope"
 end