+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* FIX: Respect default_scopes for relationships (@boblail)
+
 ## v1.0.0.rc2 (2019 Jun 17)
 
 * FEATURE: Add structured attributes to allow nesting attributes as a hash (@kobsy)